1998-12-04  Miguel de Icaza  <miguel@nuclecu.unam.mx>
	* cmd.c, tree.c: Updated to the new file.c api.

	* file.c: (copy_file_file, copy_dir_dir, move_file_file,
	move_dir_dir): Updated to work with the new updating stuff.

	* setup.c: New configuration option to compute the totals to be
	transfered.

	* file.c (compute_dir_size, panel_compute_totals): New routines to
	compute the total ammount of information that is going to be
	transfered.

+++ gnome/gdialogs.c
@@ -134,7 +134,7 @@ file_progress_show_count (long done, long total)
 }
 
 FileProgressStatus
-file_progress_show_bytes (long done, long total)
+file_progress_show_bytes (double done, double total)
 {
         if (!total)
                 gtk_progress_bar_update (GTK_PROGRESS_BAR (byte_prog), 0.0);
