+++ src/filegui.h
@@ -7,7 +7,7 @@
 FileProgressStatus file_progress_check_buttons (void);
 FileProgressStatus file_progress_show          (long done, long total);
 FileProgressStatus file_progress_show_count    (long done, long total);
-FileProgressStatus file_progress_show_bytes    (long done, long total);
+FileProgressStatus file_progress_show_bytes    (double done, double total);
 FileProgressStatus file_progress_show_source   (char *path);
 FileProgressStatus file_progress_show_target   (char *path);
 FileProgressStatus file_progress_show_deleting (char *path);
